+++ src/Schema/Directives/RelationDirective.php
@@ -32,7 +32,14 @@ abstract class RelationDirective extends BaseDirective implements FieldResolver
                 $decorateBuilder = $this->makeBuilderDecorator($resolveInfo);
                 $paginationArgs = $this->paginationArgs($args);
 
-                if (config('lighthouse.batchload_relations')) {
+                /** @var \Illuminate\Database\Eloquent\Relations\Relation $relation */
+                $relation = $parent->{$relationName}();
+
+                if (
+                    config('lighthouse.batchload_relations')
+                    // Batch loading joins across both models, thus only works if they are on the same connection
+                    && $relation->getParent()->getConnectionName() === $relation->getRelated()->getConnectionName()
+                ) {
                     /** @var \Nuwave\Lighthouse\Execution\BatchLoader\RelationBatchLoader $relationBatchLoader */
                     $relationBatchLoader = BatchLoaderRegistry::instance(
                         $this->qualifyPath($args, $resolveInfo),
@@ -47,9 +54,6 @@ abstract class RelationDirective extends BaseDirective implements FieldResolver
 
                     return $relationBatchLoader->load($parent);
                 }
-
-                /** @var \Illuminate\Database\Eloquent\Relations\Relation $relation */
-                $relation = $parent->{$relationName}();
 
                 $decorateBuilder($relation);